@@ -13,72 +13,71 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.9",
-    "@angular/cdk": "^6.4.7",
-    "@angular/common": "^6.1.9",
-    "@angular/compiler": "^6.1.9",
-    "@angular/core": "^6.1.9",
-    "@angular/flex-layout": "^6.0.0-beta.18",
-    "@angular/forms": "^6.1.9",
-    "@angular/http": "^6.1.9",
-    "@angular/material": "^6.0.0",
-    "@angular/platform-browser": "^6.1.9",
-    "@angular/platform-browser-dynamic": "^6.1.9",
-    "@angular/router": "^6.1.9",
-    "angular-pipes": "^8.0.0",
-    "angular-tree-component": "^7.2.0",
+    "@angular/animations": "^8.2.8",
+    "@angular/cdk": "^8.2.1",
+    "@angular/common": "^8.2.8",
+    "@angular/compiler": "^8.2.8",
+    "@angular/core": "^8.2.8",
+    "@angular/flex-layout": "^8.0.0-beta.27",
+    "@angular/forms": "^8.2.8",
+    "@angular/material": "^8.2.1",
+    "@angular/platform-browser": "^8.2.8",
+    "@angular/platform-browser-dynamic": "^8.2.8",
+    "@angular/router": "^8.2.8",
+    "angular-pipes": "^9.1.0",
+    "angular-tree-component": "^8.4.0",
     "bcrypt": "^3.0.2",
     "body-parser": "^1.18.2",
     "compression": "^1.7.2",
     "cookie-parser": "^1.4.3",
-    "core-js": "2.5.7",
+    "core-js": "3.2.1",
     "cors": "^2.8.4",
-    "dotenv": "^6.0.0",
+    "dotenv": "^8.1.0",
     "events": "^3.0.0",
     "express": "^4.16.3",
     "express-async-handler": "^1.1.3",
     "express-jwt": "^5.3.1",
     "express-validation": "^1.0.2",
     "formidable": "^1.2.1",
-    "helmet": "^3.12.0",
+    "helmet": "^3.21.1",
     "http-errors": "^1.6.3",
-    "joi": "^13.3.0",
+    "joi": "^14.3.1",
     "jsonwebtoken": "^8.2.1",
-    "method-override": "^2.3.10",
-    "mongoose": "^5.1.0",
+    "method-override": "^3.0.0",
+    "mongoose": "^5.7.1",
     "morgan": "^1.9.1",
-    "ng2-dragula": "^1.5.0",
+    "ng2-dragula": "^2.1.1",
     "nodemon": "^1.17.5",
     "passport": "^0.4.0",
     "passport-jwt": "^4.0.0",
     "passport-local": "^1.0.0",
-    "rxjs": "6.2.0",
+    "rxjs": "6.5.3",
     "rxjs-compat": "^6.0.0-rc.0",
-    "swagger-ui-express": "^3.0.9",
-    "webpack-dev-server": "^3.1.10",
-    "zone.js": "0.8.26"
+    "swagger-ui-express": "^4.1.1",
+    "webpack-dev-server": "^3.8.1",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.7.0-beta.1",
-    "@angular/cli": "^6.1.9",
-    "@angular/compiler-cli": "^6.1.9",
-    "@angular/language-service": "^6.1.9",
-    "@types/jasmine": "~2.8.3",
+    "@angular-devkit/build-angular": "~0.803.6",
+    "@angular/cli": "^8.3.6",
+    "@angular/compiler-cli": "^8.2.8",
+    "@angular/language-service": "^8.2.8",
+    "@types/jasmine": "~3.4.0",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~10.3.0",
-    "codelyzer": "4.3.0",
-    "concurrently": "^3.5.1",
+    "@types/node": "~12.7.8",
+    "codelyzer": "^5.0.1",
+    "concurrently": "^4.1.2",
     "del-cli": "^3.0.0",
-    "jasmine-core": "~3.1.0",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^3.1.3",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "2.0.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "1.1.0",
+    "karma": "^4.3.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "2.1.0",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "1.4.2",
     "protractor": "^5.4.1",
-    "ts-node": "~6.1.0",
-    "tslint": "~5.10.0",
-    "typescript": "~2.9.2"
+    "ts-node": "~8.4.1",
+    "tslint": "~5.20.0",
+    "typescript": "~3.5.0"
   }
 }
